+++ dt-common/src/monitor/counter_type.rs
@@ -1,7 +1,5 @@
 use strum::{Display, EnumString, IntoStaticStr};
 
-use super::task_metrics::TaskMetricValue;
-
 #[derive(EnumString, IntoStaticStr, Display, PartialEq, Eq, Hash, Clone)]
 pub enum CounterType {
     // time window counter, aggregate by: sum, avg/max/min by second
@@ -33,13 +31,15 @@ pub enum CounterType {
     #[strum(serialize = "record_size")]
     RecordSize,
 
-    // no window counter
+    // time window counter, aggregate by sample
     #[strum(serialize = "pipeline_sink_parallel_utilization")]
     PipelineSinkParallelUtilization,
     #[strum(serialize = "pipeline_sink_duration_seconds")]
     PipelineSinkDurationSeconds,
     #[strum(serialize = "partitioner_duration_seconds")]
     PartitionerDurationSeconds,
+
+    // no window counter
     #[strum(serialize = "pipeline_sink_operations_total")]
     PipelineSinkOperationsTotal,
     #[strum(serialize = "plan_records")]
@@ -88,15 +88,6 @@ pub enum WindowType {
 }
 
 impl CounterType {
-    pub fn initial_value(&self) -> TaskMetricValue {
-        match self {
-            Self::PipelineSinkParallelUtilization
-            | Self::PipelineSinkDurationSeconds
-            | Self::PartitionerDurationSeconds => TaskMetricValue::Float(0.0),
-            _ => TaskMetricValue::Integer(0),
-        }
-    }
-
     pub fn get_window_type(&self) -> WindowType {
         match self {
             Self::BatchWriteFailures
@@ -111,12 +102,12 @@ impl CounterType {
             | Self::DataBytes
             | Self::RecordSize
             | Self::ExtractedRecords
-            | Self::ExtractedBytes => WindowType::TimeWindow,
-
-            Self::PipelineSinkParallelUtilization
+            | Self::ExtractedBytes
+            | Self::PipelineSinkParallelUtilization
             | Self::PipelineSinkDurationSeconds
-            | Self::PartitionerDurationSeconds
-            | Self::PipelineSinkOperationsTotal
+            | Self::PartitionerDurationSeconds => WindowType::TimeWindow,
+
+            Self::PipelineSinkOperationsTotal
             | Self::PlanRecordTotal
             | Self::SinkedRecordTotal
             | Self::SinkedByteTotal
@@ -130,15 +121,20 @@ impl CounterType {
 
     pub fn get_aggregate_types(&self) -> Vec<AggregateType> {
         match self.get_window_type() {
-            WindowType::NoWindow => match self {
-                Self::PipelineSinkParallelUtilization => vec![AggregateType::AvgByCount],
-                Self::PipelineSinkDurationSeconds | Self::PartitionerDurationSeconds => {
-                    vec![AggregateType::Sum, AggregateType::AvgByCount]
-                }
-                _ => vec![AggregateType::Latest],
-            },
+            WindowType::NoWindow => vec![AggregateType::Latest],
 
             WindowType::TimeWindow => match self {
+                Self::PipelineSinkParallelUtilization
+                | Self::PipelineSinkDurationSeconds
+                | Self::PartitionerDurationSeconds => {
+                    vec![
+                        AggregateType::Latest,
+                        AggregateType::AvgByCount,
+                        AggregateType::MinByCount,
+                        AggregateType::MaxByCount,
+                    ]
+                }
+
                 Self::RecordsPerQuery
                 | Self::RtPerQuery
                 | Self::BufferSize